Refactoring parse module
