+++ prp/parse/typing.py
@@ -6,18 +6,12 @@ import logging
 
 from ..models.sample import MethodIndex
 from ..models.typing import (
-    LineageInformation,
-    ResultLineageBase,
-    TbProfilerLineage,
     TypingMethod,
     TypingResultCgMlst,
-    TypingResultGeneAllele,
     TypingResultMlst,
     ChewbbacaErrors,
 )
 from ..models.typing import TypingSoftware as Software
-from .phenotype.serotypefinder import parse_serotype_gene
-from .phenotype.virulencefinder import parse_vir_gene
 
 LOG = logging.getLogger(__name__)
 
@@ -156,106 +150,3 @@ def parse_cgmlst_results(
     return MethodIndex(
         type=TypingMethod.CGMLST, software=Software.CHEWBBACA, result=results
     )
-
-
-def parse_tbprofiler_lineage_results(pred_res: dict) -> MethodIndex:
-    """Parse tbprofiler results for lineage object."""
-    LOG.info("Parsing lineage results")
-    # lineages
-    lineages = [
-        LineageInformation(
-            lineage=lin["lineage"],
-            family=lin["family"],
-            rd=lin["rd"],
-            fraction=lin["fraction"],
-            support=lin["support"],
-        )
-        for lin in pred_res["lineage"]
-    ]
-    # combine into result
-    result_obj = TbProfilerLineage(
-        main_lineage=pred_res["main_lineage"],
-        sublineage=pred_res["sub_lineage"],
-        lineages=lineages,
-    )
-    # store result as a method index
-    return MethodIndex(
-        type=TypingMethod.LINEAGE,
-        software=Software.TBPROFILER,
-        result=result_obj,
-    )
-
-
-def parse_mykrobe_lineage_results(pred_res: dict) -> MethodIndex | None:
-    """Parse mykrobe results for lineage object."""
-    LOG.info("Parsing lineage results")
-    if pred_res:
-        lineage = pred_res[0]["lineage"]
-        # cast to lineage object
-        result_obj = ResultLineageBase(
-            main_lineage=lineage.split(".")[0],
-            sublineage=lineage,
-        )
-        return MethodIndex(
-            type=TypingMethod.LINEAGE, software=Software.MYKROBE, result=result_obj
-        )
-    return None
-
-
-def parse_virulencefinder_stx_typing(path: str) -> MethodIndex | None:
-    """Parse virulencefinder's output re stx typing"""
-    LOG.info("Parsing virulencefinder stx results")
-    with open(path, "rb") as inpt:
-        pred_obj = json.load(inpt)
-        # if has valid results
-        pred_result = None
-        if "virulencefinder" in pred_obj:
-            results = pred_obj["virulencefinder"]["results"]
-            species = list(results)
-            for assay, result in results[species[0]].items():
-                # skip non typing results
-                if not assay == "stx":
-                    continue
-
-                # if no stx gene was identified
-                if isinstance(result, str):
-                    continue
-
-                # take first result as the valid prediction
-                hit = next(iter(result.values()))
-                vir_gene = parse_vir_gene(hit)
-                gene = TypingResultGeneAllele(**vir_gene.model_dump())
-                pred_result = MethodIndex(
-                    type=TypingMethod.STX,
-                    software=Software.VIRULENCEFINDER,
-                    result=gene,
-                )
-    return pred_result
-
-
-def parse_serotypefinder_oh_typing(path: str) -> MethodIndex | None:
-    """Parse 's output re OH typing"""
-    LOG.info("Parsing serotypefinder oh type results")
-    with open(path, "rb") as inpt:
-        pred_obj = json.load(inpt)
-        # if has valid results
-        pred_result = []
-        if "serotypefinder" in pred_obj:
-            results = pred_obj["serotypefinder"]["results"]
-            for serotype in results:
-                # if no serotype gene was identified
-                if isinstance(results[serotype], str) or results[serotype] == {}:
-                    continue
-
-                # take first result as the valid prediction
-                hit = next(iter(results[serotype].values()))
-                vir_gene = parse_serotype_gene(hit)
-                gene = TypingResultGeneAllele(**vir_gene.model_dump())
-                pred_result.append(
-                    MethodIndex(
-                        type=serotype,
-                        software=Software.SEROTYPEFINDER,
-                        result=gene,
-                    )
-                )
-    return pred_result
