+++ prp/parse/utils.py
@@ -10,14 +10,6 @@ from ..models.phenotype import (
 )
 
 
-def _default_amr_phenotype() -> PhenotypeInfo:
-    return PhenotypeInfo(
-        type=ElementType.AMR,
-        group=ElementType.AMR,
-        name=ElementType.AMR,
-    )
-
-
 def is_prediction_result_empty(result: ElementTypeResult) -> bool:
     """Check if prediction result is emtpy.
