+++ prp/parse/virulencefinder.py
@@ -3,12 +3,17 @@ import json
 import logging
 from typing import Any
 
-from ...models.phenotype import ElementType, ElementVirulenceSubtype
-from ...models.phenotype import PredictionSoftware as Software
-from ...models.phenotype import (
+from ..models.phenotype import ElementType, ElementVirulenceSubtype
+from ..models.phenotype import PredictionSoftware as Software
+from ..models.phenotype import (
     VirulenceElementTypeResult,
     VirulenceGene,
     VirulenceMethodIndex,
+)
+from ..models.sample import MethodIndex
+from ..models.typing import (
+    TypingMethod,
+    TypingResultGeneAllele,
 )
 
 LOG = logging.getLogger(__name__)
@@ -40,7 +45,7 @@ def parse_vir_gene(
     )
 
 
-def _parse_virulencefinder_vir_results(pred: str) -> VirulenceElementTypeResult:
+def _parse_vir_results(pred: str) -> VirulenceElementTypeResult:
     """Parse virulence prediction results from virulencefinder."""
     # parse virulence finder results
     species = list(k for k in pred["virulencefinder"]["results"])
@@ -65,7 +70,7 @@ def _parse_virulencefinder_vir_results(pred: str) -> VirulenceElementTypeResult:
     return VirulenceElementTypeResult(genes=genes, phenotypes={}, variants=[])
 
 
-def parse_virulencefinder_vir_pred(path: str) -> VirulenceMethodIndex | None:
+def parse_virulence_pred(path: str) -> VirulenceMethodIndex | None:
     """Parse virulencefinder virulence prediction results.
 
     :param file: File name
@@ -77,7 +82,7 @@ def parse_virulencefinder_vir_pred(path: str) -> VirulenceMethodIndex | None:
     with open(path, "rb") as inpt:
         pred = json.load(inpt)
         if "virulencefinder" in pred:
-            results: VirulenceElementTypeResult = _parse_virulencefinder_vir_results(
+            results: VirulenceElementTypeResult = _parse_vir_results(
                 pred
             )
             result = VirulenceMethodIndex(
@@ -86,3 +91,34 @@ def parse_virulencefinder_vir_pred(path: str) -> VirulenceMethodIndex | None:
         else:
             result = None
     return result
+
+
+def parse_stx_typing(path: str) -> MethodIndex | None:
+    """Parse virulencefinder's output re stx typing"""
+    LOG.info("Parsing virulencefinder stx results")
+    with open(path, "rb") as inpt:
+        pred_obj = json.load(inpt)
+        # if has valid results
+        pred_result = None
+        if "virulencefinder" in pred_obj:
+            results = pred_obj["virulencefinder"]["results"]
+            species = list(results)
+            for assay, result in results[species[0]].items():
+                # skip non typing results
+                if not assay == "stx":
+                    continue
+
+                # if no stx gene was identified
+                if isinstance(result, str):
+                    continue
+
+                # take first result as the valid prediction
+                hit = next(iter(result.values()))
+                vir_gene = parse_vir_gene(hit)
+                gene = TypingResultGeneAllele(**vir_gene.model_dump())
+                pred_result = MethodIndex(
+                    type=TypingMethod.STX,
+                    software=Software.VIRULENCEFINDER,
+                    result=gene,
+                )
+    return pred_result
